Backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
python-dotenv>=1.0.1
pydantic-settings>=2.2.1
openai>=1.14.0
chromadb>=0.4.24
python-multipart>=0.0.9
pypdf>=4.1.0
python-docx>=1.1.0
tiktoken>=0.6.0
pymupdf>=1.23.0
pandas>=2.2.0
openpyxl>=3.1.2
pytesseract>=0.3.10
pdf2image>=1.17.0
pillow>=10.2.0

sqlalchemy>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0

# Security & Validation (New)
slowapi>=0.1.9
python-magic>=0.4.27
python-magic-bin>=0.4.14

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
